refactor: #683 BlockPropertyPanel 폼 중복 제거 및 구성 정리
